requirements.txt
streamlit
requests
beautifulsoup4
selenium
webdriver-manager
ollama
pillow
opencv-python
numpy
langchain
pyyaml
